+++ src/main/java/persistence/sql/ddl/CreateQueryBuilder.java
@@ -2,6 +2,9 @@ package persistence.sql.ddl;
 
 import static java.lang.String.format;
 
+import persistence.dialect.Dialect;
+import persistence.vender.dialect.H2Dialect;
+
 public class CreateQueryBuilder<T> extends QueryBuilder<T> {
 
     private static final String DEFAULT_CREATE_TABLE_COLUMNS_FORMAT = "(%s)";
